Move round data to model, renamed playing round to engine^
