+++ tests/game/test_round_engine.py
@@ -1,12 +1,11 @@
 import pytest
 from src.model.player_hand import Player, CardHand
 from src.model.card import Card, CardType
-from src.game.playing_round import GameRound
-
+from src.game.round_engine import RoundEngine
 
 def test_initial_round_type():
     """Test if round_type is initially None."""
-    game_round = GameRound([])
+    game_round = RoundEngine([])
     assert game_round.round_data.round_type is None
 
 
@@ -16,7 +15,7 @@ def test_player_play_card():
     john = Player("John", CardHand([rocket_2]))
 
     # Player should be able to play the card
-    game_round = GameRound(players = [john])
+    game_round = RoundEngine(players = [john])
     played_player = game_round.player_play_card(john, rocket_2)
     assert played_player == john
     assert game_round.round_data.card_by_player[john] == rocket_2
@@ -30,7 +29,7 @@ def test_player_play_card_twice():
     john = Player("John", CardHand([blue_3, blue_7]))
 
     # First card should work
-    game_round = GameRound(players = [john])
+    game_round = RoundEngine(players = [john])
     game_round.player_play_card(john, blue_3)
 
     # Second card should raise an error
@@ -42,7 +41,7 @@ def test_get_round_winner_no_cards_played():
     """Test if no winner is determined if no cards are played."""
     john = Player("John", CardHand([Card(CardType.BLUE, 3), Card(CardType.BLUE, 7)]))
 
-    game_round = GameRound(players = [john])
+    game_round = RoundEngine(players = [john])
 
     # Impossible to determine winner, as not all players played yet
     with pytest.raises(ValueError):
@@ -58,7 +57,7 @@ def test_get_round_winner_non_rocket():
     julie = Player("Julie", CardHand([blue_7]))
     matthew = Player("Matthew", CardHand([blue_5]))
 
-    game_round = GameRound(players = [john, julie, matthew])
+    game_round = RoundEngine(players = [john, julie, matthew])
 
     # Players play their cards
     game_round.player_play_card(john, blue_3)
@@ -80,7 +79,7 @@ def test_get_round_winner_rocket_wins():
     julie = Player("Julie", CardHand([blue_7]))
     matthew = Player("Matthew", CardHand([rocket_2]))
 
-    game_round = GameRound(players = [john, julie, matthew])
+    game_round = RoundEngine(players = [john, julie, matthew])
 
     # Players play their cards
     game_round.player_play_card(john, blue_3)
@@ -102,7 +101,7 @@ def test_get_round_winner_wrong_type_are_ignored():
     julie = Player("Julie", CardHand([yellow_9]))
     matthew = Player("Matthew", CardHand([pink_9]))
 
-    game_round = GameRound(players = [john, julie, matthew])
+    game_round = RoundEngine(players = [john, julie, matthew])
 
     # Players play their cards
     game_round.player_play_card(john, blue_1)
@@ -128,7 +127,7 @@ def test_get_round_winner_rocket_highest():
     anne = Player("Anne", CardHand([yellow_9]))
     luke = Player("Luke", CardHand([rocket_2]))
 
-    game_round = GameRound(players = [john, julie, matthew, anne, luke])
+    game_round = RoundEngine(players = [john, julie, matthew, anne, luke])
 
     # Players play their cards
     game_round.player_play_card(john, rocket_4)
